SVC: Move WSGI script to the application
